kismet: second chance page replacement
We will use the Second Chance page replacement algorithm for our
on-disk LRU-ish cache: the only metadata we need is already present
in standard mtime / atime file timestamps, even with relatime.

The algorithm is also well suited to batching maintenance work, and
can be invoked from time to time to amortise the directory scan.