second_chance: implement the replacement policy
We will use the Second Chance replacement algorithm for our on-disk
LRU-ish cache: the only metadata we need is already present in
standard mtime / atime file timestamps, even with relatime.

The algorithm is also well suited to batching maintenance work, and
can be invoked from time to time to amortise the directory scan.

TESTED=new unit and property-based tests.