raw_cache: an implementation of second chance on top of file timestamps
We can treat any directory of files as a second chance cache, by
ordering the second chance FIFO by modification time, and considering
that a file has been accessed since it last entered the end of the
FIFO when its access time is later than its modification time.

When the cache works well (or at least, when the caching logic shows
up on a profile), most accesses should be cache hits.  Our encoding
for second chance can rely on the operating system to naturally update
the access time when it opens a cache file, so does not add additional
operations (there will be more frequent atime updates in a relatime
system, so there is an I/O overhead).

TESTED=new unit tests (on a real filesystem).