cache_dir: move common directory maintenance logic to a trait
We will want to reuse most of the logic for the hash-sharded cache
directory.  However, we don't actually want to store any explicit
state for all the subdirectories.  Expose the state with accessors,
and implement the actual cache maintenance logic in a trait coded in
terms of these accessors.

TESTED=existing tests.