sharded: balance load with the power of two choice
Now that we have load estimates for each shard, we can ask for a
secondary hash, and use that to map each key to two potential shards,
a primary and a secondary.

For writes, we unconditionally attempt to delete the corresponding
file in the shard we did *not* select, before publishing the new value
to the least loaded shard.  On tie, we always go for the primary one.

For reads, we attempt to read from primary shard id, and then from
secondary.  Once shards have reached capacity, this should lead to
only one copy being consistently touched and thus kept alive.

TESTED=updated tests.