sharded: determinicise maintenance
We already track load estimates for each shard.  It's easy to then
compare that estimate with the target shard capacity and forcibly
trigger a maintenance when the actual load is wayy (2x) over target.

We want to let the load grow far past the load target before forcing a
maintenance because it's important for amortisation, and because we
would prefer to let the random trigger tell us when to perform
maintenance work: the jitter helps avoid synchronisation between
maintenance calls and between maintenance and other code in the same
process or in different processes that access the same file system.

TESTED=existing tests. updated smoke test with random trigger disabled.