stack: add convenience get_or_update / ensure functions
The `get_or_update` method implements the common pattern of fetching
a value from a cache and either filling the cache on miss, or otherwise
determining whether we want to use the value verbatim or update it.

When values never expire, we never want to update the value; that's
what `ensure` implements.

TESTED=new unit tests.