sharded: prefix cache subdirectory names with `.kismet_`
The dot means these internal subdirectories are guaranteed not to
collide with any file left from plain unsharded caching in the same
directory.  It's unfortunate that it also makes the files hidden by
default, but hopefully this can discourage direct uninformed
manipulation of the cache directory.

The `.kismet_` prefix also makes it easier to map directories back to
this library, and leaves the rest of the `.` namespace to the
application.

TESTED=existing tests.