stack: add new auto-sync logic, enabled by default
Aim for safer defaults, by letting the higher-level writable cache
sync files before publishing, at least by default.

Unfortunately, we sometimes have to re-open a path only to sync it,
and that's an error-prone pattern, especially if a caller retries to
publish the same file: the fsync call is liable to succeed silently.
That's why the new code panics when we fail to sync a file by path.

We can now remove the scary warning from the doc comments in `lib.rs`,
and move them to the lower level layers, `plain.rs` and `sharded.rs`.

Also edit the doc comment in `lib.rs` a little.

TESTED=updated tests.