multiplicative_hash: new module for const-fn mix + remap
Sharded caches need a deterministic function to mix potentially dodgy
hash values and map them to a range.  Steal the logic (and improve it
by fully implementing Dietzfelbinger's multiplicative hash), and make
it possible to initialise the parameters with a compile-time sha256,
instead of manually inputting pseudorandom values.

TESTED=new and existing tests.