readonly: implement optional runtime consistency check
Applications that use Kismet may want to confirm that they populate
caches in a way that makes sense.  For example, when multiple caches
have an entry for the same key, we may expect the cache values to be
identical.

Applications may now pass a consistency checker function; Kismet will
invoke that function when it finds multiple hits, and propagate any
error.