stack: also check consistency with demand-filling function
When a consistency checker is provided, `get_or_update` and `ensure`
are treated as promises that cache values are equivalent to what the
`populate` function would generate.  We can thus enforce a three-way
equality between the value returned by `populate`, any value in the
read-write cache, and any value in the read cache.