raw_cache: set atime firmly in the past when clearing accessed bit
Kismet implements a second chance cache with `atime > mtime` as
the accessed bit.  On relatime filesystems with coarse timestamp
granularity, we have to set atime in the past if we want an atime
update on access.  It's also useful to avoid surprises on relatime
filesystems with very fine access timestamps.

It's also safe to mess with atime because we don't compare atimes,
except to determine when a file has been accessed since the last
time its mtime was changed, with `atime > mtime`.  In second chance,
we want to order cache entries by their rank in the eviction list,
and that's mtime.

Fixes tests on APFS.