benign_error: small module to handle ESTALE like ENOENT
Most cases where we check for ENOENT, we want to bubble that up like
a non-erroneous cache miss. ESTALE means we thought there was something
there, but that handle (inode) has since disappeared, either because
the file was unlinked or because it was replaced with another.

Either way, it seems simplest to treat such stale filehandles like
missing files.

We don't expect this to happen a lot, except for cleanups: files in the
kismet cache should mostly be write-once (content addressed), until the
cache grows too large and old unused files are reaped to make room. In
that situation, a stale filehandle means the file was deleted, and it's
fine to treat that like ENOENT.

Alternatively, one could also flush the client-side filehandle cache
and retry, but that's more work for marginal benefits.